requirements.txt
Django==1.7.4
MySQL-python==1.2.5
Pillow==2.7.0
SQLAlchemy==0.9.8
argparse==1.2.1
dj-database-url==0.3.0
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.4.0
django-postgrespool==0.3.0
django-recaptcha==1.0.3
django-simple-captcha==0.4.4
django-simple-math-captcha==1.0.5
freetype-py==1.0
gunicorn==19.2.1
passlib==1.6.2
psycopg2==2.6
six==1.9.0
whitenoise==1.0.6
wsgiref==0.1.2
